Address boris review
